[BSP/FIXUP] remove old reset drivers for RK3500
