+++ bsp/rockchip/rk3500/driver/Kconfig
@@ -2,6 +2,5 @@ menu "RT-Thread rockchip RK3500 drivers"
 
 source "$BSP_DIR/driver/clk/Kconfig"
 source "$BSP_DIR/driver/uart8250/Kconfig"
-source "$BSP_DIR/driver/reset/Kconfig"
 source "$BSP_DIR/driver/hwtimer/Kconfig"
 endmenu
