--- bsp/rockchip/rk3500/driver/reset/Kconfig
@@ -1,5 +1,0 @@
-config RT_USING_RESET
-    bool "Using Reset Controller support"
-    depends on RT_USING_DM
-    select RT_USING_OFW
-    default n
