--- bsp/rockchip/rk3500/driver/reset/SConscript
@@ -1,13 +1,0 @@
-from building import *
-
-cwd = GetCurrentDir()
-CPPPATH = [cwd]
-
-src = []
-
-if GetDepend(['RT_USING_RESET']):
-    src += ['reset.c']
-
-group = DefineGroup('DeviceDrivers', src, depend = [''], CPPPATH = CPPPATH)
-
-Return('group')
